# openapi.yaml
openapi: 3.1.0
info:
  title: Musubi Core API
  description: The canonical HTTP surface over Musubi Core. Read + write; full v0.1
    surface.
  version: 0.1.0
paths:
...
  /v1/ops/debug/trigger-synthesis:
    post:
      tags:
      - ops
      summary: Trigger Synthesis
      description: 'Drive one synthesis tick on demand.


        Test-hook for the integration harness (Issue #119). Restricted to

        operator-scope tokens; the synthesis loop is expensive enough that

        arbitrary callers shouldn''t trigger it. Production workers run

        synthesis on their own tick loop; this endpoint exists so the

        integration suite can deterministically exercise the path without

        an embedded worker.'
      operationId: debug_trigger_synthesis.bucket=default
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerSynthesisRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerSynthesisResponse'
        '403':
          description: Caller does not hold operator scope.
        '501':
          description: '`simulate_ollama_offline=false` but no production Ollama client
            impl is wired in. See Issue #119 for context.'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/episodic/{object_id}:
    get:
      tags:
      - episodic
      summary: Get Memory
      operationId: get_memory_v1_episodic__object_id__get
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodicMemory'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - episodic-writes
      summary: Patch Episodic
      description: 'Update non-state metadata on an existing episodic row.


        State changes are forbidden on this surface — they go through

        POST /v1/lifecycle/transition so the lifecycle ledger records every

        state mutation. Tags / importance / summary are non-state metadata

        and land via a payload-only ``set_payload`` (analogous to the

        enrichment writes the maturation sweep does — see

        :mod:`musubi.lifecycle.maturation` for precedent).'
      operationId: patch_episodic.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchEpisodicRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodicMemory'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - episodic-writes
      summary: Delete Episodic
      description: 'Soft-delete by default (state → archived via the canonical

        ``transition()`` primitive). ``?hard=true`` requires operator

        scope and removes the point from Qdrant entirely.


        The operator-scope check on the hard path reads the

        :class:`AuthContext` that the outer ``require_auth`` dependency has

        already attached to ``request.state.auth``.


        Neither path deserializes the row. ``get()`` model-validates, so a row with
        an

        unmodeled payload key raised a 500 *inside the 404-guard*, before the delete
        ran —

        the row was unremovable precisely because it was broken, which is exactly
        backwards.


        **The hard path delegates to ``plane.delete()``.** It used to do its own

        ``plane.exists()`` check and then delete via an ``object_id`` PAYLOAD FILTER
        — both

        of which locate the row by payload fields. So a row that had lost or malformed
        its

        ``namespace`` / ``object_id`` keys returned 404 and stayed stored, **through
        the very

        route the fleet and operators actually use.** Rev3 hardened the SDK method
        with

        deterministic point-ID addressing and left this route on the old contract:
        the path

        nobody calls was fixed, and the path that actually failed in production was
        not.

        (Yua, rev3 review of PR #398.)


        ``plane.delete()`` addresses the point by its deterministic ID, enforces the
        stored

        namespace when the payload can state one, normalizes an unreadable prior state
        for the

        audit record, and emits the LifecycleEvent. One hardened contract, one place.


        The soft path remains payload-filtered by design: archiving requires constructing
        a

        valid lifecycle transition, which an identity-damaged row cannot support.
        Such a row

        is removable (hard delete), not archivable — and removable is what matters.'
      operationId: delete_episodic.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      - name: hard
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Hard
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '202':
          description: Transition durably pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionPendingBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/episodic:
    get:
      tags:
      - episodic
      summary: List Memories
      operationId: list_memories_v1_episodic_get
      parameters:
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_EpisodicMemory_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - episodic-writes
      summary: Capture
      operationId: capture_episodic.bucket=capture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/curated/{object_id}:
    get:
      tags:
      - curated
      summary: Get Curated
      operationId: get_curated_v1_curated__object_id__get
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuratedKnowledge'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - curated-writes
      summary: Patch Curated
      operationId: patch_curated.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCuratedRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuratedKnowledge'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - curated-writes
      summary: Delete Curated
      operationId: delete_curated.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '202':
          description: Transition durably pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionPendingBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/curated:
    get:
      tags:
      - curated
      summary: List Curated
      operationId: list_curated_v1_curated_get
      parameters:
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_CuratedKnowledge_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - curated-writes
      summary: Create Curated
      operationId: create_curated.bucket=capture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuratedCreateRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuratedCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/concepts/{object_id}:
    get:
      tags:
      - concept
      summary: Get Concept
      operationId: get_concept_v1_concepts__object_id__get
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynthesizedConcept'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - concept-writes
      summary: Delete Concept
      operationId: delete_concept.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '202':
          description: Transition durably pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionPendingBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/thoughts/stream:
    get:
      tags:
      - thoughts
      summary: Stream Thoughts
      description: 'SSE endpoint for real-time thought delivery.


        Replay: if ``Last-Event-ID: <ksuid>`` is present, every thought matching

        the subscription scope where ``object_id > last_event_id`` (lex,

        ascending) is emitted before entering live-tail mode. Capped at 500

        events — if more matched, the ``X-Musubi-Replay-Truncated: true``

        response header tells the client to fall back to ``/v1/thoughts/history``

        for deeper backfill.'
      operationId: stream_thoughts.bucket=default
      parameters:
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      - name: include
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Include
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last-Event-Id
      responses:
        '200':
          description: Server-sent event stream of delivered thoughts.
          content:
            application/json:
              schema:
                title: Response Stream Thoughts.Bucket=Default
            text/event-stream:
              schema:
                type: string
        '403':
          description: Caller does not hold read scope for `namespace`.
        '503':
          description: Thought broker unavailable.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/idempotency/receipts/lookup:
    post:
      tags:
      - idempotency
      summary: Lookup Receipt
      operationId: lookup_idempotency_receipt.bucket=default
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptLookupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptLookupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/episodic/batch:
    post:
      tags:
      - episodic-writes
      summary: Batch Capture
      operationId: batch_capture.bucket=batch-write
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCaptureRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchCaptureResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/concepts/{object_id}/reinforce:
    post:
      tags:
      - concept-writes
      summary: Reinforce Concept
      operationId: reinforce_concept.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReinforceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynthesizedConcept'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/artifacts/{object_id}/archive:
    post:
      tags:
      - artifact-writes
      summary: Archive Artifact
      operationId: archive_artifact.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '202':
          description: Transition durably pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitionPendingBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/artifacts/{object_id}/purge:
    post:
      tags:
      - artifact-writes
      summary: Purge Artifact
      description: 'Operator-only hard delete of the artifact metadata + blob.


        Reads the AuthContext attached by the outer ``require_auth`` dep

        and rejects with 403 if the operator scope isn''t on the bearer.'
      operationId: purge_artifact.bucket=default
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      - name: namespace
        in: query
        required: true
        schema:
          type: string
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/thoughts/read:
    post:
      tags:
      - thoughts-writes
      summary: Read Thoughts
      operationId: read_thoughts.bucket=default
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThoughtReadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThoughtReadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/retrieve/stream:
    post:
      tags:
      - retrieve-stream
      summary: Retrieve Stream
      operationId: retrieve_stream.bucket=default
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveQuery'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    CaptureRequest:
      properties:
        namespace:
          type: string
          title: Namespace
        content:
          type: string
          minLength: 1
          title: Content
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        tags:
          items:
            type: string
          type: array
          title: Tags
        importance:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Importance
          default: 5
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - namespace
      - content
      title: CaptureRequest
...
    EpisodicMemory:
      properties:
        object_id:
          type: string
          title: Object Id
        namespace:
          type: string
          title: Namespace
        identity_family:
          anyOf:
          - type: string
          - type: 'null'
          title: Identity Family
        schema_version:
          type: integer
          title: Schema Version
          default: 1
        created_at:
          type: string
          format: date-time
          title: Created At
        created_epoch:
          anyOf:
          - type: number
          - type: 'null'
          title: Created Epoch
        updated_at:
          type: string
          format: date-time
          title: Updated At
        updated_epoch:
          anyOf:
          - type: number
          - type: 'null'
          title: Updated Epoch
        version:
          type: integer
          title: Version
          default: 1
        state:
          type: string
          enum:
          - provisional
          - matured
          - demoted
          - archived
          - superseded
          title: State
          default: provisional
        content:
          type: string
          minLength: 1
          title: Content
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        tags:
          items:
            type: string
          type: array
          title: Tags
        importance:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Importance
          default: 5
        reinforcement_count:
          type: integer
          minimum: 0.0
          title: Reinforcement Count
          default: 0
        last_accessed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Accessed At
        access_count:
          type: integer
          minimum: 0.0
          title: Access Count
          default: 0
        supersedes:
          items:
            type: string
          type: array
          title: Supersedes
        superseded_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Superseded By
        merged_from:
          items:
            type: string
          type: array
          title: Merged From
        linked_to_topics:
          items:
            type: string
          type: array
          title: Linked To Topics
        supported_by:
          items:
            $ref: '#/components/schemas/ArtifactRef'
          type: array
          title: Supported By
        contradicts:
          items:
            type: string
          type: array
          title: Contradicts
        derived_from:
          anyOf:
          - type: string
          - type: 'null'
          title: Derived From
        valid_from:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Valid From
        valid_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Valid Until
        valid_from_epoch:
          anyOf:
          - type: number
          - type: 'null'
          title: Valid From Epoch
        valid_until_epoch:
          anyOf:
          - type: number
          - type: 'null'
          title: Valid Until Epoch
        event_at:
          type: string
          format: date-time
          title: Event At
        ingested_at:
          type: string
          format: date-time
          title: Ingested At
        modality:
          type: string
          enum:
          - text
          - voice-transcript
          - tool-call
          - system-event
          title: Modality
          default: text
        participants:
          items:
            type: string
          type: array
          title: Participants
        source_context:
          type: string
          title: Source Context
          description: Freeform origin hint, e.g. 'Claude Code session 2026-04-17
            14:23'.
          default: ''
        topics:
          items:
            type: string
          type: array
          title: Topics
        importance_last_scored_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Importance Last Scored At
      additionalProperties: false
      type: object
      required:
      - namespace
      - content
      title: EpisodicMemory
      description: 'One captured event — a message, a tool call, a system signal.


        ``event_at`` is when the thing happened in the world. ``ingested_at`` is when

        Musubi learned about it (typically == ``created_at``).'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_EpisodicMemory_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EpisodicMemory'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      title: Page[EpisodicMemory]
...
    TransitionPendingBody:
      properties:
        status:
          type: string
          const: pending
          title: Status
          default: pending
        operation_key:
          type: string
          minLength: 1
          title: Operation Key
        event_id:
          type: string
          minLength: 1
          title: Event Id
      additionalProperties: false
      type: object
      required:
      - operation_key
      - event_id
      title: TransitionPendingBody
      description: A transition accepted for reconciler completion, never fabricated
        as final.
...
    TriggerSynthesisRequest:
      properties:
        namespace:
          type: string
          title: Namespace
          description: Agent-scoped namespace prefix (e.g. 'nyla/voice'). The synthesis
            loop operates on '<prefix>/episodic' + '<prefix>/concept'.
        simulate_ollama_offline:
          type: boolean
          title: Simulate Ollama Offline
          description: Wire a stub Ollama client that always returns None (simulates
            an unreachable LLM). When false, the endpoint returns 501 — a real Ollama
            client impl is future work (there's no production OllamaClient in the
            codebase yet; the Protocol is satisfied by workers only).
          default: false
      type: object
      required:
      - namespace
      title: TriggerSynthesisRequest
      description: Debug-endpoint payload for :func:`trigger_synthesis`.
    TriggerSynthesisResponse:
      properties:
        namespace:
          type: string
          title: Namespace
        identity_family:
          type: string
          title: Identity Family
        memories_selected:
          type: integer
          title: Memories Selected
        clusters_formed:
          type: integer
          title: Clusters Formed
        concepts_created:
          type: integer
          title: Concepts Created
        concepts_reinforced:
          type: integer
          title: Concepts Reinforced
        contradictions_detected:
          type: integer
          title: Contradictions Detected
        cursor_advanced_to:
          anyOf:
          - type: number
          - type: 'null'
          title: Cursor Advanced To
        candidates_carried_forward:
          type: integer
          title: Candidates Carried Forward
          default: 0
        candidates_pruned:
          type: integer
          title: Candidates Pruned
          default: 0
      type: object
      required:
      - namespace
      - identity_family
      - memories_selected
      - clusters_formed
      - concepts_created
      - concepts_reinforced
      - contradictions_detected
      - cursor_advanced_to
      title: TriggerSynthesisResponse
      description: 'Same shape as ``musubi.lifecycle.synthesis.SynthesisReport``.


        The ``namespace`` field carries the **identity family** (e.g. ``"aoi"``)

        that synthesis actually clustered, NOT an echo of the input

        namespace. Per v1.5.5+''s per-family synthesis (musubi#335), the loop

        operates on identity-family scope; the field name is preserved for

        backward-compat with log scrapers but the semantic shifted from

        "input namespace echo" to "scope actually clustered." The new

        ``identity_family`` field below makes this explicit for new

        callers that want unambiguous semantics. See musubi#352 for the

        investigation that surfaced this naming confusion.'
...